+++ common-db/src/test/java/cn/featherfly/common/db/dialect/PostgresqlDialectTest.java
@@ -172,7 +172,7 @@ public class PostgresqlDialectTest extends DialectTest {
         String sql = dialect.buildUpsertSql("user", new String[] { "id", "name", "age" }, "id");
         System.out.println(sql);
         assertEquals(sql,
-                "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?) ON CONFLICT (id) DO UPDATE SET id=EXCLUDED.id, name=EXCLUDED.name, age=EXCLUDED.age");
+                "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?) ON CONFLICT (id) DO UPDATE SET \"name\"=EXCLUDED.\"name\", \"age\"=EXCLUDED.\"age\"");
     }
 
     /**
@@ -184,6 +184,28 @@ public class PostgresqlDialectTest extends DialectTest {
         String sql = dialect.buildUpsertBatchSql("user", new String[] { "id", "name", "age" }, "id", 2);
         System.out.println(sql);
         assertEquals(sql,
-                "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?),(?, ?, ?) ON CONFLICT (id) DO UPDATE SET id=EXCLUDED.id, name=EXCLUDED.name, age=EXCLUDED.age");
+                "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?),(?, ?, ?) ON CONFLICT (id) DO UPDATE SET \"name\"=EXCLUDED.\"name\", \"age\"=EXCLUDED.\"age\"");
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Test
+    void testInsert() {
+        String sql = dialect.buildInsertSql("user", new String[] { "id", "name", "age" });
+        System.out.println(sql);
+        assertEquals(sql, "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?)");
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Test
+    @Override
+    void testInsertBatch() {
+        String sql = dialect.buildInsertBatchSql("user", new String[] { "id", "name", "age" }, 2);
+        System.out.println(sql);
+        assertEquals(sql, "INSERT INTO \"user\" (\"id\", \"name\", \"age\") VALUES (?, ?, ?),(?, ?, ?)");
     }
 }